+++ e2e/factories/factories/factory.ts
@@ -1,13 +1,13 @@
-import type {PersistenceAdapter} from './persistence/types';
+import type {PersistenceAdapter} from '../persistence/adapter';
 
 export abstract class Factory<TOptions = unknown, TResult = unknown> {
     abstract name: string;
     abstract entityType: string; // Used for persistence routing
-    
+
     private persistence?: PersistenceAdapter;
     private trackingEnabled = true;
     protected createdEntities = new Map<string, TResult>();
-    
+
     /**
      * Set up any necessary resources (database connections, etc.)
      * Override this method if your factory needs initialization
@@ -15,7 +15,7 @@ export abstract class Factory<TOptions = unknown, TResult = unknown> {
     async setup(): Promise<void> {
         // Default implementation - no setup needed
     }
-    
+
     /**
      * Factory-specific cleanup (override if needed)
      * Called after entities are cleaned up
@@ -23,26 +23,26 @@ export abstract class Factory<TOptions = unknown, TResult = unknown> {
     async destroy(): Promise<void> {
         // Default implementation - no cleanup needed
     }
-    
+
     /**
      * Build an object without persisting it
      */
     abstract build(options?: TOptions): TResult | Promise<TResult>;
-    
+
     /**
      * Inject persistence adapter
      */
     setPersistence(adapter: PersistenceAdapter): void {
         this.persistence = adapter;
     }
-    
+
     /**
      * Control entity tracking for cleanup
      */
     setTracking(enabled: boolean): void {
         this.trackingEnabled = enabled;
     }
-    
+
     /**
      * Build and persist an object
      */
@@ -50,13 +50,13 @@ export abstract class Factory<TOptions = unknown, TResult = unknown> {
         if (!this.persistence) {
             throw new Error(`No persistence adapter configured for ${this.name} factory`);
         }
-        
+
         // Build the entity
         const entity = await Promise.resolve(this.build(options));
-        
+
         // Persist using the adapter
         const persisted = await this.persistence.insert(this.entityType, entity);
-        
+
         // Track for cleanup if enabled
         if (this.trackingEnabled) {
             const id = this.extractId(persisted);
@@ -64,10 +64,10 @@ export abstract class Factory<TOptions = unknown, TResult = unknown> {
                 this.createdEntities.set(id, persisted);
             }
         }
-        
+
         return persisted;
     }
-    
+
     /**
      * Clean up all created entities
      */
@@ -75,15 +75,15 @@ export abstract class Factory<TOptions = unknown, TResult = unknown> {
         if (!this.persistence || this.createdEntities.size === 0) {
             return;
         }
-        
+
         const ids = Array.from(this.createdEntities.keys());
         await this.persistence.deleteMany(this.entityType, ids);
         this.createdEntities.clear();
-        
+
         // Call factory-specific cleanup
         await this.destroy();
     }
-    
+
     /**
      * Extract ID from entity (override if needed)
      */
@@ -92,7 +92,7 @@ export abstract class Factory<TOptions = unknown, TResult = unknown> {
         const id = e.id || e.uuid;
         return typeof id === 'string' ? id : undefined;
     }
-    
+
     /**
      * Get all created entities
      */
