+++ e2e/factories/factories/index.ts
@@ -1,0 +1,3 @@
+export * from './post-factory';
+export * from './page-hit-factory';
+export * from './factory';
