+++ e2e/factories/factories/post-factory.ts
@@ -1,7 +1,70 @@
-import {Factory} from '../../../base-factory';
+import {Factory} from './factory';
 import {faker} from '@faker-js/faker';
-import {generateId, generateUuid, generateSlug} from '../../../utils';
-import type {PostOptions, PostResult} from './types';
+import {generateId, generateUuid, generateSlug} from '../utils/utils';
+
+// Ghost Post Options based on the Ghost schema
+export interface PostOptions {
+    id?: string;
+    uuid?: string;
+    title?: string;
+    slug?: string;
+    mobiledoc?: string | null;
+    lexical?: string | null;
+    html?: string | null;
+    comment_id?: string | null;
+    plaintext?: string | null;
+    feature_image?: string | null;
+    featured?: boolean;
+    type?: 'post' | 'page';
+    status?: 'published' | 'draft' | 'scheduled' | 'sent';
+    locale?: string | null;
+    visibility?: string;
+    email_recipient_filter?: string;
+    created_at?: Date;
+    created_by?: string;
+    updated_at?: Date;
+    updated_by?: string;
+    published_at?: Date | null;
+    published_by?: string | null;
+    custom_excerpt?: string | null;
+    codeinjection_head?: string | null;
+    codeinjection_foot?: string | null;
+    custom_template?: string | null;
+    canonical_url?: string | null;
+    newsletter_id?: string | null;
+    show_title_and_feature_image?: boolean;
+}
+
+// Return type for created posts
+export type PostResult = PostOptions & {
+    id: string;
+    uuid: string;
+    title: string;
+    slug: string;
+    status: PostOptions['status'];
+    type: PostOptions['type'];
+    created_at: Date;
+    updated_at: Date;
+    mobiledoc: string | null;
+    lexical: string | null;
+    html: string | null;
+    comment_id: string | null;
+    plaintext: string | null;
+    feature_image: string | null;
+    locale: string | null;
+    visibility: string;
+    email_recipient_filter: string;
+    created_by?: string;
+    updated_by?: string;
+    custom_excerpt: string | null;
+    codeinjection_head: string | null;
+    codeinjection_foot: string | null;
+    custom_template: string | null;
+    canonical_url: string | null;
+    newsletter_id: string | null;
+    show_title_and_feature_image: boolean;
+    featured: boolean;
+};
 
 /**
  * Simple factory for creating Ghost posts.
@@ -9,16 +72,16 @@ import type {PostOptions, PostResult} from './types';
 export class PostFactory extends Factory<PostOptions, PostResult> {
     name = 'post';
     entityType = 'posts'; // Maps to 'posts' table
-    
+
     constructor() {
         super();
     }
-    
+
     build(options: PostOptions = {}): PostResult {
         const now = new Date();
         const title = options.title || faker.lorem.sentence();
         const content = faker.lorem.paragraphs(3);
-        
+
         // Generate mobiledoc format
         const mobiledoc = {
             version: '0.3.1',
@@ -28,7 +91,7 @@ export class PostFactory extends Factory<PostOptions, PostResult> {
             sections: [[1, 'p', [[0, [], 0, content]]]],
             ghostVersion: '5.0'
         };
-        
+
         // Create defaults object
         const defaults = {
             id: generateId(),
@@ -58,7 +121,7 @@ export class PostFactory extends Factory<PostOptions, PostResult> {
             newsletter_id: null,
             show_title_and_feature_image: true
         };
-        
+
         // Merge with user options, handling special cases
         const post: PostResult = {
             ...defaults,
@@ -66,7 +129,7 @@ export class PostFactory extends Factory<PostOptions, PostResult> {
             // Handle published_at logic - if status is published but no published_at is set, use current time
             published_at: options.status === 'published' && !options.published_at ? now : (options.published_at || defaults.published_at)
         } as PostResult;
-        
+
         return post;
     }
 }
