+++ e2e/factories/plugins/ghost/database.ts
@@ -1,24 +1,20 @@
 import knex, {Knex} from 'knex';
-import {getGhostConfig} from './config';
+import {ghostConfig} from '../../config/persistence';
 
 /**
  * Simple database connection for Ghost.
  */
 export function createDatabase(): Knex {
-    const config = getGhostConfig();
     return knex({
         client: 'mysql2',
         connection: {
-            ...config,
+            ...ghostConfig(),
             charset: 'utf8mb4'
         },
         pool: {min: 0, max: 5}
     });
 }
 
-/**
- * Get site UUID from Ghost settings
- */
 export async function getSiteUuid(db: Knex): Promise<string | null> {
     try {
         const result = await db('settings')
